requirements.txt
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.1
gspread==6.0.0
oauth2client==4.1.3
flask
